FuseJS: rename class (2)
